Combined Tables/Step_4_Final/requirements.txt
google-cloud-bigquery
pandas
pyarrow
functions-framework
flask
db-dtypes  